Move range-specific translator classes into a range sub-package
