+++ sandbox/plugins/dsl-query-executor/src/main/java/org/opensearch/dsl/query/BaseTranslatorMapper.java
@@ -12,6 +12,7 @@ import org.apache.calcite.rel.type.RelDataTypeField;
 import org.apache.calcite.rex.RexNode;
 import org.opensearch.dsl.converter.ConversionContext;
 import org.opensearch.dsl.converter.ConversionException;
+import org.opensearch.dsl.query.range.BoundRequest;
 
 import java.util.Optional;
