+++ sandbox/plugins/dsl-query-executor/src/main/java/org/opensearch/dsl/query/DefaultTranslatorMapper.java
@@ -15,6 +15,7 @@ import org.apache.calcite.sql.fun.SqlStdOperatorTable;
 import org.apache.calcite.sql.type.SqlTypeName;
 import org.opensearch.dsl.converter.ConversionContext;
 import org.opensearch.dsl.converter.ConversionException;
+import org.opensearch.dsl.query.range.RangeBoundMath;
 
 import java.util.Optional;
