+++ sandbox/plugins/dsl-query-executor/src/main/java/org/opensearch/dsl/query/TimestampTranslatorMapper.java
@@ -17,6 +17,8 @@ import org.apache.calcite.sql.type.SqlTypeName;
 import org.apache.calcite.util.TimestampString;
 import org.opensearch.dsl.converter.ConversionContext;
 import org.opensearch.dsl.converter.ConversionException;
+import org.opensearch.dsl.query.range.BoundRequest;
+import org.opensearch.dsl.query.range.RangeDateParsing;
 
 import java.util.Optional;
