+++ sandbox/plugins/dsl-query-executor/src/main/java/org/opensearch/dsl/query/range/RangeBoundMath.java
@@ -6,19 +6,19 @@
  * compatible open source license.
  */
 
-package org.opensearch.dsl.query;
+package org.opensearch.dsl.query.range;
 
 import org.apache.calcite.sql.type.SqlTypeName;
 
 /**
  * Generic numeric predicates and type-narrowing helpers shared by the range bound translation path.
  */
-final class RangeBoundMath {
+public final class RangeBoundMath {
 
     private RangeBoundMath() {}
 
     /** Returns true if the SqlTypeName represents an integer-family type (not float/double/decimal). */
-    static boolean isIntegerType(SqlTypeName typeName) {
+    public static boolean isIntegerType(SqlTypeName typeName) {
         return typeName == SqlTypeName.INTEGER
             || typeName == SqlTypeName.BIGINT
             || typeName == SqlTypeName.SMALLINT
@@ -29,7 +29,7 @@ final class RangeBoundMath {
      * Returns true if the numeric value has a non-zero fractional part.
      * Mirrors legacy NumberFieldMapper.hasDecimalPart.
      */
-    static boolean hasDecimalPart(Object value) {
+    public static boolean hasDecimalPart(Object value) {
         if (value instanceof Number) {
             double d = ((Number) value).doubleValue();
             return d % 1 != 0;
@@ -41,7 +41,7 @@ final class RangeBoundMath {
      * Returns the signum (-1, 0, or 1) of a numeric value.
      * Mirrors legacy NumberFieldMapper.signum.
      */
-    static double signum(Object value) {
+    public static double signum(Object value) {
         if (value instanceof Number) {
             return Math.signum(((Number) value).doubleValue());
         }
@@ -52,7 +52,7 @@ final class RangeBoundMath {
      * Truncates a numeric value to long (floor toward zero), supporting all integer family widths.
      * Used as the base truncation before narrowing to the specific integer type.
      */
-    static long toLongValue(Object value) {
+    public static long toLongValue(Object value) {
         if (value instanceof Number) {
             return ((Number) value).longValue();
         }
@@ -63,7 +63,7 @@ final class RangeBoundMath {
      * Narrows a long value to the appropriate Java type for the given SqlTypeName.
      * INTEGER/SMALLINT/TINYINT produce Integer; BIGINT produces Long.
      */
-    static Number narrowToFieldType(long value, SqlTypeName typeName) {
+    public static Number narrowToFieldType(long value, SqlTypeName typeName) {
         if (typeName == SqlTypeName.BIGINT) {
             return value;
         }
@@ -74,7 +74,7 @@ final class RangeBoundMath {
      * Returns the maximum value for the given integer-family SqlTypeName.
      * Used for overflow guard checks before incrementing truncated values.
      */
-    static long getMaxValueForType(SqlTypeName typeName) {
+    public static long getMaxValueForType(SqlTypeName typeName) {
         switch (typeName) {
             case BIGINT:
                 return Long.MAX_VALUE;
@@ -93,7 +93,7 @@ final class RangeBoundMath {
      * Returns the minimum value for the given integer-family SqlTypeName.
      * Used for overflow guard checks before decrementing truncated values.
      */
-    static long getMinValueForType(SqlTypeName typeName) {
+    public static long getMinValueForType(SqlTypeName typeName) {
         switch (typeName) {
             case BIGINT:
                 return Long.MIN_VALUE;
