+++ sandbox/plugins/dsl-query-executor/src/main/java/org/opensearch/dsl/query/range/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+/** Range query translator and supporting bound-processing helpers. */
+package org.opensearch.dsl.query.range;
