+++ sandbox/plugins/dsl-query-executor/src/test/java/org/opensearch/dsl/query/range/RangeQueryTranslatorTests.java
@@ -6,7 +6,7 @@
  * compatible open source license.
  */
 
-package org.opensearch.dsl.query;
+package org.opensearch.dsl.query.range;
 
 import org.apache.calcite.rex.RexCall;
 import org.apache.calcite.rex.RexInputRef;
